Renamed SanWebSocket as AnalysisWebSocket
