+++ assets/js/pages/database/event_stats/index.js
@@ -1,5 +1,5 @@
 import { eventStatsModal } from './EventStatsModal.js';
-import { sanWebSocket } from '../../../SanWebSocket.js';
+import { sanWebSocket } from '../../../AnalysisWebSocket.js';
 
 await sanWebSocket.connect();
