+++ assets/js/pages/learn/analysis/index.js
@@ -1,6 +1,6 @@
 import { FEN } from '@chesslablab/chessboard';
 import { gameForm } from './GameForm.js';
-import { sanWebSocket } from '../../../SanWebSocket.js';
+import { sanWebSocket } from '../../../AnalysisWebSocket.js';
 
 await sanWebSocket.connect();
