+++ assets/js/pages/openings/OpeningsTable.js
@@ -1,6 +1,6 @@
 import { Movetext } from '@chesslablab/jsblab';
 import AbstractComponent from '../../AbstractComponent.js';
-import { sanWebSocket } from '../../SanWebSocket.js';
+import { sanWebSocket } from '../../AnalysisWebSocket.js';
 import * as mode from '../../../mode.js';
 import * as variant from '../../../variant.js';
