+++ assets/js/pages/openings/name/index.js
@@ -1,5 +1,5 @@
 import { openingsNameModal } from './OpeningsNameModal.js';
-import { sanWebSocket } from '../../../SanWebSocket.js';
+import { sanWebSocket } from '../../../AnalysisWebSocket.js';
 
 await sanWebSocket.connect();
