+++ assets/js/pages/openings/san_movetext/index.js
@@ -1,5 +1,5 @@
 import { openingsSanMovetextModal } from './OpeningsSanMovetextModal.js';
-import { sanWebSocket } from '../../../SanWebSocket.js';
+import { sanWebSocket } from '../../../AnalysisWebSocket.js';
 
 await sanWebSocket.connect();
