+++ assets/js/translations/messages.de.js
@@ -14,7 +14,7 @@ const messagesDe = {
   "Draw offer accepted": "Remisangebot akzeptiert",
   "The game is resigned": "Die Partie wurde aufgegeben",
   "Your opponent is gone": "Dein Gegner ist verschwunden",
-  // SanWebSocket
+  // AnalysisWebSocket
   "Invalid SAN movetext, please try again": "Ungültiger SAN-Zugtext, bitte erneut versuchen"
 };
